backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asttokens==2.4.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
decorator==5.1.1
distro==1.9.0
executing==2.1.0
ffmpeg-python==0.2.0
Flask==3.0.3
Flask-Cors==5.0.0
future==1.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
MarkupSafe==3.0.1
matplotlib-inline==0.1.7
numpy==2.1.2
openai==1.51.2
opencv-python==4.10.0.84
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
PyAudio==0.2.14
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tqdm==4.66.5
traitlets==5.14.3
twelvelabs==0.2.8
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.4
wheel==0.44.0
